server: manifestIDFromValues
- Also move code that turns query strings to IDs to own file.
